refactor: dropdown 리스트 가져오는 로직 분리
